WIP - Model -> Models, move all models to same namespace
